refactor: restore shared hooks file and use imports for code reuse
